fix: resolve misplaced audit findings by separating responsibilities and relocating code
Three structural issues caused audit findings: resolve.ts mixed CLI
parsing with config validation, audit/pipeline.ts owned shared rebuild
logic, and duplicates.ts had a stale residual hash.

- Rename resolve.ts → parse-flags.ts, functions → parseFlags/parseAuditFlags
- Move constraint clamping to config/validate.ts (clampAudits), keeping
  validation responsibility in the module that owns constraint knowledge
- Move loadProjectConfig call to command.ts (the orchestrator)
- Move rebuildParseResult and index helpers from audit/pipeline.ts to
  db/rebuild.ts where kural place confirmed it belongs (61% confidence)
- Update @kuralBorrows role text to cover dial parsing, snapshot-to-engine
  flow, and finding rendering
- Update stale @kuralResidual hash on duplicates.ts
