+++ src/shell/commands/audit/KURAL.md
@@ -1,2 +1,2 @@
 The inspector's office. Wires the diagnostic CLI arguments to the issue display pipeline. It is the only command that reads and renders localized issues — no other command queries snapshot data for structural diagnostics.
-@kuralBorrows analysis/audits "terminal surface that formats and renders engine output as categorized finding reports"
+@kuralBorrows analysis/audits "terminal surface that parses dial flags, opens stored snapshots to run the detection engine, and renders findings as categorized reports"
