+++ src/shell/commands/audit/parse-flags.test.ts
@@ -1,9 +1,8 @@
-import { clampOrWarn, parseOr, resolveAuditConfig, resolveConfig } from "./resolve.ts";
-import { describe, expect, it, vi } from "vite-plus/test";
+import { parseAuditFlags, parseFlags, parseOr } from "./parse-flags.ts";
+import { describe, expect, it } from "vite-plus/test";
 
 const ZERO = 0;
 const ONE = 1;
-const NEG_ONE = -1;
 const MOCK_SENSITIVITY = 1.5;
 
 describe("parseOr", () => {
@@ -44,40 +43,17 @@ describe("parseOr", () => {
   });
 });
 
-describe("clampOrWarn", () => {
-  it("returns value when valid", () => {
-    const VALUE = 2.5;
-    expect(clampOrWarn("test", VALUE, true, ZERO)).toBe(VALUE);
-  });
-
-  it("returns fallback when invalid", () => {
-    const spy = vi.spyOn(console, "error").mockImplementation(() => {});
-    const FALLBACK = 2.0;
-    expect(clampOrWarn("sensitivity", NEG_ONE, false, FALLBACK)).toBe(FALLBACK);
-    expect(spy).toHaveBeenCalledOnce();
-    spy.mockRestore();
-  });
-
-  it("logs a warning with the field name and value", () => {
-    const spy = vi.spyOn(console, "error").mockImplementation(() => {});
-    const FALLBACK = 4;
-    clampOrWarn("minGroup", ZERO, false, FALLBACK);
-    expect(spy).toHaveBeenCalledWith(expect.stringContaining("minGroup"));
-    spy.mockRestore();
-  });
-});
-
-describe("resolveConfig", () => {
+describe("parseFlags", () => {
   it("uses CLI values over project config", () => {
     const PROJECT_SENS = 1.5;
     const CLI_SENS = 3.0;
-    const result = resolveConfig({ sensitivity: "3.0" }, { sensitivity: PROJECT_SENS });
+    const result = parseFlags({ sensitivity: "3.0" }, { sensitivity: PROJECT_SENS });
     expect(result.sensitivity).toBe(CLI_SENS);
   });
 
   it("falls back to project config when CLI is absent", () => {
     const PROJECT_SENS = 1.5;
-    const result = resolveConfig({}, { sensitivity: PROJECT_SENS });
+    const result = parseFlags({}, { sensitivity: PROJECT_SENS });
     expect(result.sensitivity).toBe(PROJECT_SENS);
   });
 
@@ -85,63 +61,47 @@ describe("resolveConfig", () => {
     const DEFAULT_SENSITIVITY = 2.0;
     const DEFAULT_FLOOR = 0.9;
     const DEFAULT_GROUP = 4;
-    const result = resolveConfig({}, {});
+    const result = parseFlags({}, {});
     expect(result.sensitivity).toBe(DEFAULT_SENSITIVITY);
     expect(result.containmentFloor).toBe(DEFAULT_FLOOR);
     expect(result.minGroup).toBe(DEFAULT_GROUP);
   });
 
-  it("clamps negative sensitivity to default", () => {
-    const spy = vi.spyOn(console, "error").mockImplementation(() => {});
-    const DEFAULT_SENSITIVITY = 2.0;
-    const result = resolveConfig({ sensitivity: "-1" }, {});
-    expect(result.sensitivity).toBe(DEFAULT_SENSITIVITY);
-    spy.mockRestore();
-  });
-
-  it("clamps containmentFloor above 1 to default", () => {
-    const spy = vi.spyOn(console, "error").mockImplementation(() => {});
-    const DEFAULT_FLOOR = 0.9;
-    const result = resolveConfig({ containmentFloor: "1.5" }, {});
-    expect(result.containmentFloor).toBe(DEFAULT_FLOOR);
-    spy.mockRestore();
+  it("passes through negative sensitivity without clamping", () => {
+    const result = parseFlags({ sensitivity: "-1" }, {});
+    expect(result.sensitivity).toBe(-ONE);
   });
 
   it("accepts containmentFloor of 0", () => {
-    const result = resolveConfig({ containmentFloor: "0" }, {});
+    const result = parseFlags({ containmentFloor: "0" }, {});
     expect(result.containmentFloor).toBe(ZERO);
   });
 
   it("truncates fractional minGroup via parseInt", () => {
     const TRUNCATED = 2;
-    const result = resolveConfig({ minGroup: "2.5" }, {});
+    const result = parseFlags({ minGroup: "2.5" }, {});
     expect(result.minGroup).toBe(TRUNCATED);
   });
 
   it("passes through disable from project config", () => {
-    const result = resolveConfig({}, { disable: ["outliers"] });
+    const result = parseFlags({}, { disable: ["outliers"] });
     expect(result.disable).toEqual(["outliers"]);
   });
 });
 
-vi.mock("../../config/loader.ts", () => ({
-  loadProjectConfig: (): { audits: { sensitivity: number } } => ({
-    audits: { sensitivity: MOCK_SENSITIVITY },
-  }),
-}));
-
-describe("resolveAuditConfig", () => {
+describe("parseAuditFlags", () => {
   it("merges CLI disable with config disable", () => {
-    const { config, disabledAudits } = resolveAuditConfig({
-      disable: "outliers,bloated-files",
-    });
+    const { config, disabledAudits } = parseAuditFlags(
+      { disable: "outliers,bloated-files" },
+      { sensitivity: MOCK_SENSITIVITY },
+    );
     expect(config.sensitivity).toBe(MOCK_SENSITIVITY);
     expect(disabledAudits.has("outliers")).toBe(true);
     expect(disabledAudits.has("bloated-files")).toBe(true);
   });
 
   it("returns empty disabled set when no disable specified", () => {
-    const { disabledAudits } = resolveAuditConfig({});
+    const { disabledAudits } = parseAuditFlags({}, {});
     expect(disabledAudits.size).toBe(ZERO);
   });
 });
