+++ src/shell/config/validate.test.ts
@@ -1,5 +1,5 @@
+import { clampAudits, validateConfig } from "./validate.ts";
 import { describe, expect, it } from "vite-plus/test";
-import { validateConfig } from "./validate.ts";
 
 const ZERO = 0;
 const ONE = 1;
@@ -13,6 +13,87 @@ function auditField(config: Record<string, unknown>, key: string): unknown {
   }
   return Object.getOwnPropertyDescriptor(audits, key)?.value;
 }
+
+const DEFAULT_SENSITIVITY = 2.0;
+const DEFAULT_FLOOR = 0.9;
+const DEFAULT_GROUP = 4;
+const VALID_SENSITIVITY = 1.5;
+const VALID_FLOOR = 0.8;
+const VALID_GROUP = 5;
+
+/** Builds a valid baseline config for clamp tests. */
+function validConfig(): {
+  sensitivity: number;
+  containmentFloor: number;
+  minGroup: number;
+} {
+  return {
+    sensitivity: VALID_SENSITIVITY,
+    containmentFloor: VALID_FLOOR,
+    minGroup: VALID_GROUP,
+  };
+}
+
+describe("clampAudits", () => {
+  it("keeps valid values unchanged", () => {
+    const { config, warnings } = clampAudits(validConfig());
+    expect(warnings).toHaveLength(ZERO);
+    expect(config.sensitivity).toBe(VALID_SENSITIVITY);
+    expect(config.containmentFloor).toBe(VALID_FLOOR);
+    expect(config.minGroup).toBe(VALID_GROUP);
+  });
+
+  it("clamps negative sensitivity to default", () => {
+    const { config, warnings } = clampAudits({ ...validConfig(), sensitivity: -ONE });
+    expect(config.sensitivity).toBe(DEFAULT_SENSITIVITY);
+    expect(warnings).toHaveLength(ONE);
+    expect(warnings[ZERO]).toContain("sensitivity must be positive");
+  });
+
+  it("clamps zero sensitivity to default", () => {
+    const { config, warnings } = clampAudits({ ...validConfig(), sensitivity: ZERO });
+    expect(config.sensitivity).toBe(DEFAULT_SENSITIVITY);
+    expect(warnings).toHaveLength(ONE);
+  });
+
+  it("clamps containmentFloor above 1 to default", () => {
+    const OVER = 1.5;
+    const { config, warnings } = clampAudits({ ...validConfig(), containmentFloor: OVER });
+    expect(config.containmentFloor).toBe(DEFAULT_FLOOR);
+    expect(warnings[ZERO]).toContain("containmentFloor must be between 0 and 1");
+  });
+
+  it("clamps negative containmentFloor to default", () => {
+    const NEG = -0.1;
+    const { config, warnings } = clampAudits({ ...validConfig(), containmentFloor: NEG });
+    expect(config.containmentFloor).toBe(DEFAULT_FLOOR);
+    expect(warnings).toHaveLength(ONE);
+  });
+
+  it("accepts containmentFloor of 0", () => {
+    const { config, warnings } = clampAudits({ ...validConfig(), containmentFloor: ZERO });
+    expect(config.containmentFloor).toBe(ZERO);
+    expect(warnings).toHaveLength(ZERO);
+  });
+
+  it("clamps fractional minGroup to default", () => {
+    const FRAC = 2.5;
+    const { config, warnings } = clampAudits({ ...validConfig(), minGroup: FRAC });
+    expect(config.minGroup).toBe(DEFAULT_GROUP);
+    expect(warnings[ZERO]).toContain("minGroup must be a positive integer");
+  });
+
+  it("clamps zero minGroup to default", () => {
+    const { config, warnings } = clampAudits({ ...validConfig(), minGroup: ZERO });
+    expect(config.minGroup).toBe(DEFAULT_GROUP);
+    expect(warnings).toHaveLength(ONE);
+  });
+
+  it("preserves disable field", () => {
+    const { config } = clampAudits({ ...validConfig(), disable: ["outliers"] });
+    expect(config.disable).toEqual(["outliers"]);
+  });
+});
 
 describe("validateConfig — non-object roots", () => {
   it("rejects null", () => {
